Add a getter to get a random game
feat: Add a getter to get a random game

This commit adds a new getter to the store module for games, which returns a random game object from the list of games. The getter is called 'randomGame' and is used in the 'RecommendedGames' component to display a random game recommendation.

The getter uses the 'Math.random()' function to generate a random index number between 0 and the length of the game list array, and then returns the game object at that index.

This feature enhances the user experience by providing them with a random game recommendation each time they visit the 'featuredGame' component.